refactored component for adding routing
