+++ src/Details.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+class Details extends Component {
+    state = {  }
+    render() { 
+        return ( 
+            <div>
+            <h1>Hi lol</h1>
+            </div>
+         );
+    }
+}
+ 
+export default Details;
